requirements.txt
Flask>=3.0.0
flask_cors
mysql-connector-python>=8.2.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0